requirements.txt
python-dotenv
langchain
langchain_openai